made requested updates to PR
